Rename builder -> doc package, fix printer resolution
