+++ src/common/fast-path.js
