+++ src/doc/index.js
@@ -1,0 +1,7 @@
+"use strict";
+
+module.exports = {
+  builders: require("./doc-builders"),
+  printer: require("./doc-printer"),
+  utils: require("./doc-utils")
+};
