+++ src/language-graphql/printer-graphql.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const docBuilders = require("../builder/doc-builders");
+const docBuilders = require("../doc").builders;
 const concat = docBuilders.concat;
 const join = docBuilders.join;
 const hardline = docBuilders.hardline;
