+++ src/main/parser.js
@@ -9,16 +9,37 @@ const parsers = loadPlugins().reduce(
   {}
 );
 
-function resolveParseFunction(opts) {
+// Copy the "parse" function from parser to a new object whose values are
+// functions. Use defineProperty()/getOwnPropertyDescriptor() such that we
+// don't invoke the parser.parse getters.
+const parsersForCustomParserApi = Object.keys(parsers).reduce(
+  (object, parserName) =>
+    Object.defineProperty(
+      object,
+      parserName,
+      Object.getOwnPropertyDescriptor(parsers[parserName], "parse")
+    ),
+  {}
+);
+
+function resolveParser(opts) {
   if (typeof opts.parser === "function") {
-    return opts.parser;
+    // Custom parser API always works with JavaScript.
+    return {
+      parse: opts.parser,
+      astFormat: "estree"
+    };
   }
+
   if (typeof opts.parser === "string") {
     if (parsers.hasOwnProperty(opts.parser)) {
-      return parsers[opts.parser].parse;
+      return parsers[opts.parser];
     }
     try {
-      return eval("require")(path.resolve(process.cwd(), opts.parser));
+      return {
+        parse: eval("require")(path.resolve(process.cwd(), opts.parser)),
+        astFormat: "estree"
+      };
     } catch (err) {
       /* istanbul ignore next */
       throw new ConfigError(`Couldn't resolve parser "${opts.parser}"`);
@@ -29,10 +50,10 @@ function resolveParseFunction(opts) {
 }
 
 function parse(text, opts) {
-  const parseFunction = resolveParseFunction(opts);
+  const parser = resolveParser(opts);
 
   try {
-    return parseFunction(text, parsers, opts);
+    return parser.parse(text, parsersForCustomParserApi, opts);
   } catch (error) {
     const loc = error.loc;
 
@@ -50,4 +71,4 @@ function parse(text, opts) {
   }
 }
 
-module.exports = { parse };
+module.exports = { parse, resolveParser };
